Restore recoverable Host_Error and fix refactor regressions
Host_Error and Host_EndGame now throw HostAbort, which Host_Frame catches,
instead of calling Sys_Error and exiting the process. The original engine used
longjmp for this; without it every QuakeC runtime error, lost connection, or
malformed server message terminated the client. The interpreter's call stack
is reset on abort and on progs load. -fno-exceptions is gone accordingly.

Concrete fixes:
- status command passed an eastl::array and a string_view to %s (garbage or
  crash); the function is rewritten with plain control flow.
- CL_ParseUpdate lost the original colormap-index bounds check; a bad server
  could index past cl.scores. Restored as a recoverable Host_Error.
- Every copy of a network-supplied string into a fixed hostcache/name buffer
  is now bounded (MSG_ReadString returns up to 2047 bytes; the buffers hold 16).
- Warning cleanup: memset/memcpy on non-trivial QuakeC field blocks, signed/
  unsigned item-bit compares in the HUD, misleading single-line if/while
  statements, a trigraph in a format string, an exception-specifier mismatch.

Regression baselines are unchanged.

+++ src/ui/hud.cpp
@@ -346,7 +346,7 @@ void Sbar_DrawInventory()
     float time;
     int flashon;
     if (rogue) {
-        if (cl.stats[STAT_ACTIVEWEAPON] >= RIT_LAVA_NAILGUN) {
+        if (static_cast<uint32_t>(cl.stats[STAT_ACTIVEWEAPON]) >= RIT_LAVA_NAILGUN) {
             Sbar_DrawPic(0, -24, rsb_invbar[0]);
         } else {
             Sbar_DrawPic(0, -24, rsb_invbar[1]);
@@ -360,7 +360,7 @@ void Sbar_DrawInventory()
             time = cl.item_gettime[i];
             flashon = static_cast<int>((cl.time - time) * 10);
             if (flashon >= 10) {
-                if (cl.stats[STAT_ACTIVEWEAPON] == (IT_SHOTGUN << i)) {
+                if (static_cast<uint32_t>(cl.stats[STAT_ACTIVEWEAPON]) == (IT_SHOTGUN << i)) {
                     flashon = 1;
                 } else {
                     flashon = 0;
@@ -420,9 +420,9 @@ void Sbar_DrawInventory()
     }
     if (rogue) {
         // check for powered up weapon.
-        if (cl.stats[STAT_ACTIVEWEAPON] >= RIT_LAVA_NAILGUN) {
+        if (static_cast<uint32_t>(cl.stats[STAT_ACTIVEWEAPON]) >= RIT_LAVA_NAILGUN) {
             for (int i = 0; i < 5; i++) {
-                if (cl.stats[STAT_ACTIVEWEAPON] == (RIT_LAVA_NAILGUN << i)) {
+                if (static_cast<uint32_t>(cl.stats[STAT_ACTIVEWEAPON]) == (RIT_LAVA_NAILGUN << i)) {
                     Sbar_DrawPic((i + 2) * 24, -16, rsb_weapons[i]);
                 }
             }
